Start implement URI normalization: Cross Platform Edition
Making this work correctly on both CLJ and CLJS is turning out to be quite an
exercise. JavaScript lacks support for basic encoding/decoding of strings, but
goog.crypt came to the rescue. Neither JavsScript or Java correctly deals with
unicode surrogate pairs, so e.g. emojis are reported as being two characters
wide instead of one, with all due consequences.